# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RESERVA DE AULAS PARA EXAMENES ",
        "version": "1.0.0"
    },
    "paths": {
...
        "/docente/verificar/{idDocente}": {
            "put": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Verificacion de una cuenta de docente",
                "operationId": "f8aa788532d319426cc0d3fd09e42103",
                "parameters": [
                    {
                        "name": "idDocente",
                        "in": "path",
                        "description": "Id de Docente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/docente/notificaciones/{idDocente}": {
            "get": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Obtencion de las notificaciones de un docente apartir de un id",
                "operationId": "1f1a5994870ee069ced9f19b95a8f184",
                "parameters": [
                    {
                        "name": "idDocente",
                        "in": "path",
                        "description": "Id de Docente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/docente/materias/{idDocente}": {
            "get": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Obtencion de las materias y grupos de un docente",
                "operationId": "7d392d6c52abcf74d8b0470f560de310",
                "parameters": [
                    {
                        "name": "idDocente",
                        "in": "path",
                        "description": "Id de Docente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/docente/cuentas": {
            "get": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Obtencion de las cuentas que estan solicitdas para ser aceptadas",
                "operationId": "9e3c71603a4cdfa2f8145c34a00a916e",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/docente/rechazar/{idDocente}": {
            "put": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Rechazo de la creacion de una cuenta",
                "operationId": "01e7860b984a2022664dcd7840f10199",
                "parameters": [
                    {
                        "name": "idDocente",
                        "in": "path",
                        "description": "Id de Docente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/docente/sin-registrar": {
            "get": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Obtencion de los docentes sin tegistrar",
                "operationId": "384ab9a87852f02d1a1b31f6eb0510f2",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/docente/actualizar/{idDocente}": {
            "put": {
                "tags": [
                    "Docentes"
                ],
                "summary": "Activacion de la cuenta de un docente",
                "operationId": "570e45e34b8d5b570d4c68809cbd7438",
                "parameters": [
                    {
                        "name": "idDocente",
                        "in": "path",
                        "description": "Id de Docente",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "celular": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "contrasenia": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
        "/materia/docentes/{idMateria}": {
            "get": {
                "tags": [
                    "Materia"
                ],
                "summary": "Obtencion de una todos los docentes de una materia",
                "operationId": "6a852c0c6a9fcf25bc0bc62987b91852",
                "parameters": [
                    {
                        "name": "idMateria",
                        "in": "path",
                        "description": "Id de la Materia",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/materia/grupos/{idMateria}": {
            "get": {
                "tags": [
                    "Materia"
                ],
                "summary": "Obtencion de una todos los grupos apartir de un id de materia",
                "operationId": "89733dd7699472256a4a0fa8b7ff3677",
                "parameters": [
                    {
                        "name": "idMateria",
                        "in": "path",
                        "description": "Id de la Materia",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/materia/{idMateria}": {
            "get": {
                "tags": [
                    "Materia"
                ],
                "summary": "Obtencion de una materia apartir de un id de materia",
                "operationId": "346b7942211034a885943d4beda15a57",
                "parameters": [
                    {
                        "name": "idMateria",
                        "in": "path",
                        "description": "Id de la Materia",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/materia": {
            "get": {
                "tags": [
                    "Materia"
                ],
                "summary": "Obtencion de todas las materias",
                "operationId": "5078ff9e647c0d6071ee6c12b98c462a",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}